RW-5338 delete unneeded code after enableCohortBuilderV2 flag removal
