+++ ui/src/app/cohort-search/attributes-page/attributes-page.component.spec.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import {AttributesPage} from './attributes-page.component';
 
 
-describe('AttributesPage', () => {
+describe('AttributesPageV2', () => {
   it('should render', () => {
     const wrapper = shallow(<AttributesPage node={{}} close={() => {}}/>);
     expect(wrapper.exists()).toBeTruthy();
