+++ ui/src/app/cohort-search/attributes-page/attributes-page.component.tsx
@@ -1,34 +1,35 @@
+import * as fp from 'lodash/fp';
 import {Dropdown} from 'primereact/dropdown';
 import * as React from 'react';
 
 import {PM_UNITS, PREDEFINED_ATTRIBUTES} from 'app/cohort-search/constant';
-import {ppiQuestions} from 'app/cohort-search/search-state.service';
+import {ppiQuestions, ppiSurveys} from 'app/cohort-search/search-state.service';
+import {Selection} from 'app/cohort-search/selection-list/selection-list.component';
+import {COPE_SURVEY_GROUP_NAME} from 'app/cohort-search/tree-node/tree-node.component';
 import {
   mapParameter,
   sanitizeNumericalInput,
   stripHtml,
   subTypeToTitle
 } from 'app/cohort-search/utils';
-import {Button} from 'app/components/buttons';
+import {Button, StyledAnchorTag} from 'app/components/buttons';
+import {FlexRowWrap} from 'app/components/flex';
 import {ClrIcon} from 'app/components/icons';
 import {CheckBox, NumberInput} from 'app/components/inputs';
 import {Spinner, SpinnerOverlay} from 'app/components/spinners';
 import {cohortBuilderApi} from 'app/services/swagger-fetch-clients';
 import colors, {colorWithWhiteness} from 'app/styles/colors';
-import {reactStyles, withCurrentWorkspace} from 'app/utils';
+import {reactStyles, withCurrentCohortCriteria, withCurrentWorkspace} from 'app/utils';
 import {triggerEvent} from 'app/utils/analytics';
-import {currentWorkspaceStore} from 'app/utils/navigation';
+import {currentCohortCriteriaStore, currentWorkspaceStore} from 'app/utils/navigation';
 import {WorkspaceData} from 'app/utils/workspace-data';
 import {AttrName, CriteriaSubType, Domain, Operator} from 'generated/fetch';
 
 const styles = reactStyles({
   countPreview: {
-    backgroundColor: '#E4F3FC',
-    padding: '0.5rem',
+    borderTop: `1px solid ${colorWithWhiteness(colors.black, 0.59)}`,
     marginTop: '1rem',
-    position: 'absolute',
-    width: '93%',
-    bottom: '1rem',
+    paddingTop: '1rem',
   },
   row: {
     display: 'flex',
@@ -36,33 +37,52 @@ const styles = reactStyles({
   },
   label: {
     color: colors.primary,
-    padding: '0.5rem',
-    fontWeight: 500,
+    fontWeight: 600,
     display: 'flex',
+    lineHeight: '0.75rem',
   },
   orCircle: {
-    backgroundColor: '#e2e2e9',
+    backgroundColor: colorWithWhiteness(colors.primary, 0.75),
     borderRadius: '50%',
+    color: colors.primary,
     width: '2.25rem',
     height: '2.25rem',
-    margin: '0.75rem 3rem 0.25rem',
+    margin: '0.5rem 0',
     lineHeight: '2.25rem',
     textAlign: 'center',
-    fontSize: '0.45rem',
+    fontSize: '12px',
+    fontWeight: 600
+  },
+  andCircle: {
+    backgroundColor: colorWithWhiteness(colors.primary, 0.75),
+    borderRadius: '50%',
+    color: colors.primary,
+    width: '2.25rem',
+    height: '2.25rem',
+    lineHeight: '2.25rem',
+    textAlign: 'center',
+    fontSize: '12px',
+    fontWeight: 600,
+    position: 'absolute',
+    left: '42%'
+  },
+  andDivider: {
+    height: '1.15rem',
+    marginBottom: '1.15rem',
+    width: '100%',
+    borderBottom: `2px solid ${colorWithWhiteness(colors.primary, 0.75)}`
   },
   container: {
     display: 'flex',
     marginLeft: 'auto',
     marginRight: 'auto',
-    padding: '0.2rem 0.5rem 0',
   },
   dropdown: {
     width: '12rem',
     marginRight: '1rem',
   },
   categorical: {
-    width: '25%',
-    float: 'left',
+    width: '100%',
     marginBottom: '0.25rem'
   },
   badge: {
@@ -71,22 +91,29 @@ const styles = reactStyles({
     fontSize: '10px',
     height: '0.625rem',
     padding: '0 4px',
-    marginRight: '0.5rem',
+    marginLeft: '0.25rem',
     borderRadius: '10px',
     display: 'inline-flex',
     verticalAlign: 'middle',
     alignItems: 'center',
   },
-  buttonContainer: {
-    flex: '0 0 25%',
-    maxWidth: '25%',
-    padding: '0 0.5rem',
+  addButtonContainer: {
+    bottom: '1rem',
+    position: 'absolute',
+    right: '1rem'
   },
-  button: {
+  addButton: {
+    height: '2rem',
+    borderRadius: '5px',
+    fontWeight: 600,
+    marginRight: '0.5rem'
+  },
+  calculateButton: {
     height: '1.75rem',
-    margin: '.25rem .5rem .25rem 0',
-    borderRadius: '4px',
-    fontWeight: 600
+    border: `1px solid`,
+    borderColor: colors.accent,
+    borderRadius: '2px',
+    fontWeight: 100
   },
   spinner: {
     marginRight: '0.25rem',
@@ -95,10 +122,7 @@ const styles = reactStyles({
   resultsContainer: {
     flex: '0 0 50%',
     maxWidth: '50%',
-    padding: '0.3rem 0.5rem 0',
     color: colors.primary,
-    fontSize: '14px',
-    lineHeight: '20px'
   },
   error: {
     background: colors.warning,
@@ -108,8 +132,7 @@ const styles = reactStyles({
     textAlign: 'left',
     border: '1px solid #ebafa6',
     borderRadius: '5px',
-    margin: '0.25rem 0.5rem',
-    padding: '8px',
+    margin: '0.25rem 0',
   },
   errors: {
     background: colorWithWhiteness(colors.danger, .7),
@@ -117,13 +140,68 @@ const styles = reactStyles({
     fontSize: '11px',
     border: '1px solid #ebafa6',
     borderRadius: '3px',
-    margin: '0.25rem 0.5rem',
+    margin: '0.25rem 0',
     padding: '3px 5px'
   },
   errorItem: {
     lineHeight: '16px',
   },
+  moreInfo: {
+    color: colors.accent,
+    cursor: 'pointer',
+    fontWeight: 300,
+    margin: '0.75rem 0',
+    textDecoration: 'underline'
+  }
 });
+
+interface CalculateFooterProps {
+  addButtonText: string;
+  addFn: () => void;
+  backFn: () => void;
+  calculateFn: () => void;
+  calculating: boolean;
+  count: number;
+  disableAdd: boolean;
+  disableCalculate: boolean;
+}
+
+export const CalculateFooter = (props: CalculateFooterProps) => {
+  const {addButtonText, addFn, backFn, calculateFn, calculating, count, disableAdd, disableCalculate} = props;
+  return <div style={{background: colorWithWhiteness(colors.primary, .87), bottom: 0, position: 'sticky'}}>
+    <FlexRowWrap style={styles.countPreview}>
+      <div style={styles.resultsContainer}>
+        <Button id='attributes-calculate'
+          type='secondaryLight'
+          disabled={disableCalculate}
+          style={disableCalculate
+            ? {...styles.calculateButton, borderColor: colorWithWhiteness(colors.dark, 0.6)}
+            : styles.calculateButton}
+          onClick={() => calculateFn()}>
+          {calculating && <Spinner size={16} style={styles.spinner}/>} Calculate
+        </Button>
+      </div>
+      <div style={styles.resultsContainer}>
+        <div style={{fontWeight: 600}}>Number of Participants:
+          <span> {count === null ? '--' : count.toLocaleString()} </span>
+        </div>
+      </div>
+    </FlexRowWrap>
+    <FlexRowWrap style={{flexDirection: 'row-reverse', marginTop: '0.5rem'}}>
+      <Button type='primary'
+              disabled={disableAdd}
+              style={styles.addButton}
+              onClick={() => addFn()}>
+        {addButtonText}
+      </Button>
+      <Button type='link'
+              style={{color: colors.primary, marginRight: '0.75rem'}}
+              onClick={() => backFn()}>
+        BACK
+      </Button>
+    </FlexRowWrap>
+  </div>;
+};
 
 const optionUtil = {
   ANY: {display: 'Any value', code: 'Any'},
@@ -134,15 +212,17 @@ const optionUtil = {
 };
 
 interface AttributeForm {
-  exists: boolean; // Check if attribute values exist (Measurements only)
-  num: Array<any>; // Numerical attributes (Physical Measurements or Measurements)
+  anyValue: boolean; // Include any values that exist (Measurements and COPE only)
+  anyVersion: boolean; // Include any version that exist (COPE only)
+  num: Array<any>; // Numerical attributes (Physical Measurements, Measurements or COPE)
   cat: Array<any>; // Categorical attributes (Measurements only)
 }
 
 interface Props {
+  back: Function;
   close: Function;
+  criteria: Array<Selection>;
   node: any;
-  select: Function;
   workspace: WorkspaceData;
 }
 
@@ -151,10 +231,13 @@ interface State {
   count: number;
   countError: boolean;
   form: AttributeForm;
+  formErrors: Array<string>;
+  formValid: boolean;
+  isCOPESurvey: boolean;
   loading: boolean;
   options: any;
 }
-export const AttributesPage = withCurrentWorkspace() (
+export const AttributesPage = fp.flow(withCurrentWorkspace(), withCurrentCohortCriteria()) (
   class extends React.Component<Props, State> {
     constructor(props: Props) {
       super(props);
@@ -162,7 +245,10 @@ export const AttributesPage = withCurrentWorkspace() (
         calculating: false,
         count: null,
         countError: false,
-        form: {exists: false, num: [], cat: []},
+        form: {anyValue: false, anyVersion: false, num: [], cat: []},
+        formErrors: [],
+        formValid: false,
+        isCOPESurvey: false,
         loading: true,
         options: [
           {label: 'Equals', value: Operator.EQUAL},
@@ -174,18 +260,88 @@ export const AttributesPage = withCurrentWorkspace() (
     }
 
     componentDidMount() {
-      const {node: {subtype}} = this.props;
-      const{form, options} = this.state;
-      if (!this.isMeasurement) {
-        options.unshift({label: optionUtil.ANY.display, value: AttrName[AttrName.ANY]});
+      this.initAttributeForm();
+    }
+
+    componentDidUpdate(prevProps: Readonly<Props>): void {
+      if (this.props.node !== prevProps.node) {
+        // A different node has been selected, so we reset the form and load the new attributes
+        this.setState({
+          form: {anyValue: false, anyVersion: false, num: [], cat: []},
+          formErrors: [],
+          formValid: this.isPhysicalMeasurement,
+          loading: true
+        }, () => this.initAttributeForm());
       }
-      if (this.hasRange) {
+    }
+
+    initAttributeForm() {
+      const {node: {subtype}} = this.props;
+      const {form, options} = this.state;
+      if (this.isSurvey) {
+        this.getSurveyAttributes();
+      } else if (this.isMeasurement) {
         this.getAttributes();
       } else {
+        if (!options.find(opt => opt.value === AttrName.ANY.toString())) {
+          options.unshift({label: optionUtil.ANY.display, value: AttrName[AttrName.ANY]});
+        }
         form.num = subtype === CriteriaSubType[CriteriaSubType.BP]
           ? JSON.parse(JSON.stringify(PREDEFINED_ATTRIBUTES.BP_DETAIL))
           : [{name: subtype, operator: 'ANY', operands: []}];
-        this.setState({form, options, count: this.nodeCount, loading: false});
+        this.setState({count: this.nodeCount, form, formValid: true, loading: false, options});
+      }
+    }
+
+    async getSurveyAttributes() {
+      const {node: {conceptId, parentId, path, subtype, value}} = this.props;
+      const {form, options} = this.state;
+      const {cdrVersionId} = currentWorkspaceStore.getValue();
+      const surveyId = path.split('.')[0];
+      const surveyNode = !!ppiSurveys.getValue()[cdrVersionId] && ppiSurveys.getValue()[cdrVersionId].find(n => n.id === +surveyId);
+      if (!!surveyNode && surveyNode.name === COPE_SURVEY_GROUP_NAME) {
+        const promises = [];
+        if (subtype === CriteriaSubType.QUESTION || (subtype === CriteriaSubType.ANSWER && !value)) {
+          promises.push(cohortBuilderApi().findSurveyVersionByQuestionConceptId(+cdrVersionId, surveyNode.conceptId, conceptId));
+        } else {
+          promises.push(
+            cohortBuilderApi().findSurveyVersionByQuestionConceptIdAndAnswerConceptId(
+              +cdrVersionId,
+              surveyNode.conceptId,
+              ppiQuestions.getValue()[parentId].conceptId,
+              +value
+            )
+          );
+        }
+        if (subtype === CriteriaSubType.ANSWER) {
+          promises.push(cohortBuilderApi().findCriteriaAttributeByConceptId(+cdrVersionId, conceptId));
+        }
+        const [surveyVersions, numericalAttributes] = await Promise.all(promises);
+        form.cat = surveyVersions.items.map(attr => ({
+          checked: false,
+          conceptName: attr.displayName,
+          estCount: attr.itemCount,
+          valueAsConceptId: attr.surveyVersionConceptId
+        }));
+        if (numericalAttributes && !(subtype === CriteriaSubType.ANSWER.toString() && !!value)) {
+          numericalAttributes.items.forEach(attr => {
+            if (!form.num.length) {
+              form.num.push({
+                name: AttrName.NUM,
+                operator: null,
+                operands: [],
+                conceptId: +value,
+                [attr.conceptName]: parseInt(attr.estCount, 10)
+              });
+            } else {
+              form.num[0][attr.conceptName] = parseInt(attr.estCount, 10);
+            }
+          });
+        }
+        this.setState({count: null, form, isCOPESurvey: true, loading: false});
+      } else {
+        options.unshift({label: optionUtil.ANY.display, value: AttrName[AttrName.ANY]});
+        this.setState({formValid: true, isCOPESurvey: false, options}, () => this.getAttributes());
       }
     }
 
@@ -196,6 +352,7 @@ export const AttributesPage = withCurrentWorkspace() (
       cohortBuilderApi().findCriteriaAttributeByConceptId(+cdrVersionId, conceptId).then(resp => {
         resp.items.forEach(attr => {
           if (attr.type === AttrName[AttrName.NUM]) {
+            // NUM attributes set the min and max range for the number inputs in the attributes form
             if (!form.num.length) {
               form.num.push({
                 name: AttrName.NUM,
@@ -208,30 +365,38 @@ export const AttributesPage = withCurrentWorkspace() (
               form.num[0][attr.conceptName] = parseInt(attr.estCount, 10);
             }
           } else {
+            // CAT attributes are displayed as checkboxes in the attributes form
             if (parseInt(attr.estCount, 10) > 0) {
               attr['checked'] = false;
               form.cat.push(attr);
             }
           }
         });
-        const count = this.isSurvey ? this.nodeCount : null;
-        this.setState({count, form, loading: false});
+        this.setState({count: null, form, loading: false});
       });
     }
 
-    toggleCheckbox(checked: boolean) {
+    toggleAnyValueCheckbox(checked: boolean) {
       const {form} = this.state;
       let {node: {count}} = this.props;
+      form.anyValue = checked;
       if (checked) {
-        form.exists = true;
-        form.num = form.num.map(attr =>
-          ({...attr, operator: this.isPhysicalMeasurement ? 'ANY' : null, operands: []}));
-        form.cat = form.cat.map(attr => ({...attr, checked: false}));
-      } else {
-        count = null;
-        form.exists = false;
+        form.num = form.num.map(attr => ({...attr, operator: this.isPhysicalMeasurement ? 'ANY' : null, operands: []}));
+        if (this.isMeasurement) {
+          form.cat = form.cat.map(attr => ({...attr, checked: false}));
+        }
       }
-      this.setState({form, count});
+      if (!checked || count === -1) {
+        count = null;
+      }
+      this.setState({form, count}, () => this.validateForm());
+    }
+
+    toggleAnyVersionCheckbox(checked: boolean) {
+      const {form} = this.state;
+      form.anyVersion = checked;
+      form.cat = form.cat.map(attr => ({...attr, checked: false}));
+      this.setState({form}, () => this.validateForm());
     }
 
     selectChange(attributeIndex: number, value: string) {
@@ -253,65 +418,72 @@ export const AttributesPage = withCurrentWorkspace() (
         // delete second operand if it exists
         form.num[attributeIndex].operands.splice(1);
       }
-      const count = value === 'ANY' ? this.nodeCount : null;
-      this.setState({form, count});
+      const count = value === 'ANY' && !this.isSurvey ? this.nodeCount : null;
+      this.setState({form, count}, () => this.validateForm());
     }
 
     inputChange(input: string, attributeIndex: number, operandIndex: number) {
       const {form} = this.state;
       form.num[attributeIndex].operands[operandIndex] = sanitizeNumericalInput(input);
-      this.setState({form, count: null});
+      this.setState({form, count: null}, () => this.validateForm());
     }
 
     checkboxChange(checked: boolean, index: number) {
       const {form} = this.state;
       form.cat[index].checked = checked;
-      this.setState({form});
+      this.setState({form, count: null}, () => this.validateForm());
     }
 
     validateForm() {
-      const {form} = this.state;
-      let formErrors = new Set(), formValid = true, operatorSelected = true;
-      if (form.exists) {
-        return {formValid, formErrors};
+      const {form, isCOPESurvey} = this.state;
+      if ((form.anyValue || (isCOPESurvey && form.num.length === 0)) && (!isCOPESurvey || form.anyVersion)) {
+        this.setState({formValid: true, formErrors: []});
+      } else {
+        let formValid = true, operatorSelected = form.num.length !== 0;
+        const formErrors = form.num.reduce((acc, attr) => {
+          const {MIN, MAX, operator} = attr;
+          const operands = attr.operands.map(op => parseInt(op, 10));
+          switch (operator) {
+            case null:
+              operatorSelected = false;
+              return acc;
+            case 'ANY':
+              return acc;
+            case Operator.BETWEEN:
+              if (operands.length < 2) {
+                formValid = false;
+              }
+              break;
+            default:
+              if (operands.length === 0) {
+                formValid = false;
+              }
+          }
+          if (operands.includes(NaN)) {
+            formValid = false;
+            acc.add('Form can only accept valid numbers');
+          }
+          if (this.isPhysicalMeasurement && operands.some(op => op < 0)) {
+            formValid = false;
+            acc.add('Form cannot accept negative values');
+          }
+          if (this.hasRange && operands.some(op => op < MIN || op > MAX)) {
+            formValid = false;
+            acc.add(`Values must be between ${MIN.toLocaleString()} and ${MAX.toLocaleString()}`);
+          }
+          return acc;
+        }, new Set());
+        // The second condition sets formValid to false if this is a Measurements or COPE attribute with no operator selected from the
+        // dropdown and no categorical checkboxes checked
+        if (this.isMeasurement && formValid) {
+          formValid = operatorSelected || form.cat.some(attr => attr.checked);
+        }
+        if (isCOPESurvey && formValid) {
+          formValid = (form.num.length === 0 || form.anyValue || operatorSelected)
+            && (form.anyVersion || form.cat.some(attr => attr.checked));
+        }
+        this.setState({formErrors: Array.from(formErrors), formValid});
       }
-      formErrors = form.num.reduce((acc, attr) => {
-        const {MIN, MAX, operator} = attr;
-        const operands = attr.operands.map(op => parseInt(op, 10));
-        switch (operator) {
-          case null:
-            operatorSelected = false;
-            return acc;
-          case 'ANY':
-            return acc;
-          case Operator.BETWEEN:
-            if (operands.length < 2) {
-              formValid = false;
-            }
-            break;
-          default:
-            if (operands.length === 0) {
-              formValid = false;
-            }
-        }
-        if (operands.includes(NaN)) {
-          formValid = false;
-          acc.add('Form can only accept valid numbers');
-        }
-        if (this.isPhysicalMeasurement && operands.some(op => op < 0)) {
-          formValid = false;
-          acc.add('Form cannot accept negative values');
-        }
-        if (this.hasRange && operands.some(op => op < MIN || op > MAX)) {
-          formValid = false;
-          acc.add(`Values must be between ${MIN.toLocaleString()} and ${MAX.toLocaleString()}`);
-        }
-        return acc;
-      }, formErrors);
-      // The second condition sets formValid to false if this is a Measurements attribute with no operator selected from the dropdown and
-      // no categorical checkboxes checked
-      formValid = formValid && !(this.isMeasurement && !operatorSelected && !form.cat.some(attr => attr.checked));
-      return {formErrors, formValid};
     }
 
     get nodeCount() {
@@ -325,15 +497,18 @@ export const AttributesPage = withCurrentWorkspace() (
     }
 
     get paramId() {
-      const {node: {conceptId, id}} = this.props;
-      const {form} = this.state;
-      const code = form.exists ? 'Any' : form.num.reduce((acc, attr) => {
+      const {node: {conceptId, id, value}} = this.props;
+      const {form, isCOPESurvey} = this.state;
+      const code = form.anyValue ? 'Any' : form.num.reduce((acc, attr) => {
         if (attr.operator) {
           acc += optionUtil[attr.operator].code;
         }
         return acc;
       }, '');
-      return `param${(conceptId || id) + code}`;
+      const paramConceptId = isCOPESurvey && !!value ? value : conceptId;
+      // make sure param ID is unique for different checkbox combinations
+      const catValues = form.cat.filter(c => c.checked).map(c => c.valueAsConceptId).join('');
+      return `param${(paramConceptId || id) + code + catValues}`;
     }
 
     get displayName() {
@@ -342,12 +517,14 @@ export const AttributesPage = withCurrentWorkspace() (
     }
 
     get paramWithAttributes() {
-      const {node, node: {name, subtype}} = this.props;
-      const {form} = this.state;
+      const {node, node: {name, subtype, value}} = this.props;
+      const {form, isCOPESurvey} = this.state;
       let paramName;
       const attrs = [];
-      if (form.exists) {
+      if (!isCOPESurvey && form.anyValue) {
         paramName = name + ` (${optionUtil.ANY.display})`;
+      } else if (isCOPESurvey && form.anyValue && form.anyVersion) {
+        paramName = name + ` (Any version AND any value)`;
       } else {
         form.num.filter(at => at.operator).forEach(({operator, operands, conceptId}) => {
           const attr = {name: AttrName.NUM, operator, operands};
@@ -370,9 +547,31 @@ export const AttributesPage = withCurrentWorkspace() (
             }
             return checked;
           }, []);
-          attrs.push({name: AttrName.CAT, operator: Operator.IN, operands: catOperands});
+          if (isCOPESurvey && !form.anyVersion) {
+            attrs.push({
+              name: AttrName.SURVEYVERSIONCONCEPTID,
+              operator: Operator.IN,
+              operands: catOperands
+            });
+          } else if (!isCOPESurvey) {
+            attrs.push({
+              name: AttrName.CAT,
+              operator: Operator.IN,
+              operands: catOperands
+            });
+          }
         }
         paramName = this.paramName;
+      }
+      if (isCOPESurvey && subtype === CriteriaSubType.ANSWER && !!value) {
+        attrs.push({
+          name: AttrName.CAT,
+          operator: Operator.IN,
+          operands: [value]
+        });
+      }
+      if (subtype === CriteriaSubType.ANSWER && (form.anyValue || (form.num.length && form.num[0].operator === 'ANY')) && value === '') {
+        attrs.push({name: AttrName.ANY});
       }
       return {...node, parameterId: this.paramId, name: paramName, attributes: attrs};
     }
@@ -382,6 +581,9 @@ export const AttributesPage = withCurrentWorkspace() (
       const {form} = this.state;
       const selectionDisplay = [];
       let name = '';
+      if (form.anyVersion) {
+        selectionDisplay.push('Any version');
+      }
       form.num.filter(at => at.operator).forEach((attr, i) => {
         if (attr.operator === 'ANY') {
           if (i === 0) {
@@ -401,7 +603,9 @@ export const AttributesPage = withCurrentWorkspace() (
         selectionDisplay.push(name);
       }
       form.cat.filter(ca => ca.checked).forEach(attr => selectionDisplay.push(attr.conceptName));
-      const nodeName = this.isSurvey ? ppiQuestions.getValue()[node.parentId].name : node.name;
+      const nodeName = node.domainId === Domain.SURVEY && !node.group
+        ? `${ppiQuestions.getValue()[node.parentId].name} - ${node.name}`
+        : node.name;
       return nodeName + ' (' + selectionDisplay.join(', ') +
         (this.hasUnits && form.num[0].operator !== AttrName.ANY ? ' ' + PM_UNITS[node.subtype] : '') + ')';
     }
@@ -434,13 +638,15 @@ export const AttributesPage = withCurrentWorkspace() (
     }
 
     addParameterToSearchItem() {
-      const {close, select} = this.props;
+      const {close} = this.props;
+      let {criteria} = this.props;
       const param = this.paramWithAttributes;
       // TODO remove condition to only track PM criteria for 'Phase 2' of CB Google Analytics
       if (this.isPhysicalMeasurement) {
         this.trackEvent(param.subtype, 'Add');
       }
-      select(param);
+      criteria = criteria.filter(crit => crit.parameterId !== param.parameterId);
+      currentCohortCriteriaStore.next([...criteria, param]);
       close();
     }
 
@@ -481,94 +687,153 @@ export const AttributesPage = withCurrentWorkspace() (
       return this.isMeasurement || this.isSurvey;
     }
 
+    get disableAddButton() {
+      const {calculating, formValid} = this.state;
+      return calculating || !formValid;
+    }
+
+    get disableCalculateButton() {
+      const {calculating, count, form, formValid, isCOPESurvey} = this.state;
+      return calculating || !formValid
+        || (form.anyValue && count !== null)
+        || (isCOPESurvey && !form.anyVersion && !form.cat.some(attr => attr.checked))
+        || (!this.isSurvey && form.num.length && form.num.every(attr => attr.operator === 'ANY'));
+    }
+
+    renderNumericalAttributes() {
+      const {node: {count, subtype}} = this.props;
+      const {form, isCOPESurvey, options} = this.state;
+      return form.num.length > 0 && <React.Fragment>
+        {this.isMeasurement && <div style={styles.label}>Numeric Values</div>}
+        {isCOPESurvey && <div>
+          <CheckBox onChange={(v) => this.toggleAnyValueCheckbox(v)}/> Any value
+          {count > -1 && <span style={styles.badge}> {count.toLocaleString()}</span>}
+        </div>}
+        {!(isCOPESurvey && form.anyValue) && form.num.map((attr, a) => <div key={a}>
+          {this.isBloodPressure && <div style={styles.label}>{attr.name}</div>}
+          {isCOPESurvey && <div style={styles.orCircle}>OR</div>}
+          <Dropdown style={{marginBottom: '0.5rem', width: '100%'}}
+                    value={attr.operator}
+                    options={options}
+                    placeholder='Select Operator'
+                    onChange={(e) => this.selectChange(a, e.value)}/>
+          <FlexRowWrap>
+            {![null, 'ANY'].includes(attr.operator) && <div style={{width: '33%'}}>
+              <NumberInput style={{padding: '0 0.25rem', ...(this.hasUnits ? {width: '70%'} : {})}}
+                           value={attr.operands[0] || ''}
+                           min={attr.MIN} max={attr.MAX}
+                           onChange={(v) => this.inputChange(v, a, 0)}/>
+              {this.hasUnits && <span> {PM_UNITS[subtype]}</span>}
+            </div>}
+            {attr.operator === Operator.BETWEEN && <React.Fragment>
+              <div style={{padding: '0.2rem 1.5rem 0 1rem'}}>and</div>
+              <div style={{width: '33%'}}>
+                <NumberInput style={{padding: '0 0.25rem', ...(this.hasUnits ? {width: '70%'} : {})}}
+                             value={attr.operands[1] || ''}
+                             min={attr.MIN} max={attr.MAX}
+                             onChange={(v) => this.inputChange(v, a, 1)}/>
+                {this.hasUnits && <span> {PM_UNITS[subtype]}</span>}
+              </div>
+            </React.Fragment>}
+          </FlexRowWrap>
+          {this.hasRange && ![null, 'ANY'].includes(attr.operator) && <div style={{paddingTop: '0.2rem'}}>
+            Range: {attr.MIN.toLocaleString()} - {attr.MAX.toLocaleString()}
+          </div>}
+        </div>)}
+      </React.Fragment>;
+    }
+
+    renderCategoricalAttributes() {
+      const {node: {count}} = this.props;
+      const {form, isCOPESurvey} = this.state;
+      return form.cat.length > 0 && <React.Fragment>
+        {isCOPESurvey && <div>
+          <CheckBox onChange={(v) => this.toggleAnyVersionCheckbox(v)}/> Any version
+          {count > -1 && <span style={styles.badge}>{count.toLocaleString()}</span>}
+        </div>}
+        {!(isCOPESurvey && form.anyVersion) && <React.Fragment>
+          <div style={styles.orCircle}>OR</div>
+          {!isCOPESurvey && <div style={styles.label}>Categorical Values</div>}
+          {form.cat.map((attr, a) => <div key={a} style={styles.categorical}>
+            <CheckBox checked={attr.checked} style={{marginRight: '3px'}}
+                      onChange={(v) => this.checkboxChange(v, a)}/>
+            {attr.conceptName}
+            <span style={styles.badge}>{parseInt(attr.estCount, 10).toLocaleString()}</span>
+          </div>)}
+        </React.Fragment>}
+      </React.Fragment>;
+    }
+
     render() {
-      const {node} = this.props;
-      const {calculating, count, countError, form, loading, options} = this.state;
-      const {formErrors, formValid} = this.validateForm();
-      const disableAdd = calculating || !formValid;
-      const disableCalculate = disableAdd || form.exists || form.num.every(attr => attr.operator === 'ANY');
+      const {back, node: {domainId, name, parentId, subtype}} = this.props;
+      const {calculating, count, countError, form, formErrors, isCOPESurvey, loading} = this.state;
       return (loading ?
         <SpinnerOverlay/> :
-        <div style={{margin: '0.5rem 0 1.5rem'}}>
+        <div id='attributes-form' style={{marginTop: '0.5rem'}}>
+          {isCOPESurvey ? <div>
+            <h3 style={{fontWeight: 600, margin: '0 0 0.5rem', textTransform: 'capitalize'}}>
+              COPE Survey (COVID-19) attribute
+            </h3>
+            <div style={{lineHeight: '0.75rem', paddingRight: '1.5rem'}}>
+              The COPE survey is longitudinal and will change over time. Use the following attributes to select data from one or more
+               versions.
+            </div>
+            <div style={styles.moreInfo}>
+              <StyledAnchorTag
+                href='https://www.researchallofus.org/data-tools/survey-explorer/cope-survey/'
+                target='_blank'
+                rel='noopener noreferrer'>
+                More info
+              </StyledAnchorTag>
+            </div>
+          </div> : <h3 style={{fontWeight: 600, margin: '0 0 0.5rem', textTransform: 'capitalize'}}>
+            {this.isPhysicalMeasurement ? name : domainId.toString().toLowerCase()} Detail
+          </h3>}
+          {this.isSurvey && <div style={{...styles.label, marginBottom: '0.5rem'}}>
+            {subtype === CriteriaSubType.ANSWER
+              ? `${ppiQuestions.getValue()[parentId].name} - ${name}`
+              : name
+            }
+          </div>}
           {countError && <div style={styles.error}>
             <ClrIcon style={{margin: '0 0.5rem 0 0.25rem'}} className='is-solid'
               shape='exclamation-triangle' size='22'/>
             Sorry, the request cannot be completed.
           </div>}
-          {!!formErrors.size && <div style={styles.errors}>
-            {Array.from(formErrors).map((err, e) => <div key={e} style={styles.errorItem}>
+          {formErrors.length > 0 && <div style={styles.errors}>
+            {formErrors.map((err, e) => <div key={e} style={styles.errorItem}>
               {err}
             </div>)}
           </div>}
-          {this.isMeasurement && <div>
-            <div style={styles.label}>{this.displayName}</div>
-            <CheckBox style={{marginLeft: '0.5rem'}}
-              onChange={(v) => this.toggleCheckbox(v)}/> Any value (lab exists)
-            {!form.exists && form.num.length > 0 && <div style={styles.orCircle}>OR</div>}
-          </div>}
-          {!form.exists && <React.Fragment>
-            {form.num.map((attr, a) => <div key={a}>
-              {this.isMeasurement && <div style={styles.label}>Numeric Values</div>}
-              {this.isSurvey && <div style={styles.label}>{ppiQuestions.getValue()[node.parentId].name}</div>}
-              {this.isBloodPressure && <div style={styles.label}>{attr.name}</div>}
-              <div style={styles.container}>
-                <div style={styles.dropdown}>
-                  <Dropdown style={{width: '100%'}} value={attr.operator} options={options}
-                    placeholder='Select Operator' onChange={(e) => this.selectChange(a, e.value)}/>
-                </div>
-                {![null, 'ANY'].includes(attr.operator) && <div>
-                  <NumberInput style={{padding: '0 0.25rem', width: '3rem'}}
-                    value={attr.operands[0] || ''}
-                    min={attr.MIN} max={attr.MAX}
-                    onChange={(v) => this.inputChange(v, a, 0)}/>
-                  {this.hasUnits && <span> {PM_UNITS[node.subtype]}</span>}
-                </div>}
-                {attr.operator === Operator.BETWEEN && <React.Fragment>
-                  <div style={{padding: '0.2rem 1.5rem 0 1rem'}}>and</div>
-                  <div>
-                    <NumberInput style={{padding: '0 0.25rem', width: '3rem'}}
-                      value={attr.operands[1] || ''}
-                      min={attr.MIN} max={attr.MAX}
-                      onChange={(v) => this.inputChange(v, a, 1)}/>
-                    {this.hasUnits && <span> {PM_UNITS[node.subtype]}</span>}
-                  </div>
-                </React.Fragment>}
-                {this.hasRange && ![null, 'ANY'].includes(attr.operator) &&
-                  <span style={{paddingTop: '0.2rem'}}>&nbsp;Range: {attr.MIN.toLocaleString()} - {attr.MAX.toLocaleString()}</span>
-                }
-              </div>
-            </div>)}
-            {form.cat.length > 0 && <React.Fragment>
-              <div style={styles.orCircle}>OR</div>
-              <div style={styles.label}>Categorical Values</div>
-              <div style={{marginLeft: '0.5rem'}}>
-                {form.cat.map((attr, a) => <div key={a} style={styles.categorical}>
-                  <CheckBox checked={attr.checked} style={{marginRight: '3px'}}
-                    onChange={(v) => this.checkboxChange(v, a)} />
-                  {attr.conceptName}&nbsp;
-                  <span style={styles.badge}> {parseInt(attr.estCount, 10).toLocaleString()}</span>
-                </div>)}
-              </div>
-            </React.Fragment>}
-          </React.Fragment>}
-          <div style={styles.countPreview}>
-            <div style={styles.row}>
-              <div style={styles.buttonContainer}>
-                <Button type='primary' disabled={disableCalculate} style={styles.button} onClick={() => this.requestPreview()}>
-                  {calculating && <Spinner size={16} style={styles.spinner}/>} Calculate
-                </Button>
-              </div>
-              <div style={styles.resultsContainer}>
-                <div style={{fontWeight: 'bold'}}>Results</div>
-                <div>Number Participants: <span> {count === null ? '--' : count.toLocaleString()} </span></div>
-              </div>
-              <div style={styles.buttonContainer}>
-                <Button type='primary' disabled={disableAdd} style={styles.button} onClick={() => this.addParameterToSearchItem()}>
-                  ADD THIS
-                </Button>
-              </div>
+          {isCOPESurvey
+            ? <div>
+              {this.renderCategoricalAttributes()}
+              {form.num.length > 0 && form.cat.length > 0 && <div style={{position: 'relative'}}>
+                <div style={styles.andCircle}>AND</div>
+                <div style={styles.andDivider}/>
+              </div>}
+              {this.renderNumericalAttributes()}
             </div>
-          </div>
+            : <div>
+              {this.isMeasurement && <div>
+                <div style={styles.label}>{this.displayName}</div>
+                <CheckBox onChange={(v) => this.toggleAnyValueCheckbox(v)}/> Any value (lab exists)
+                {!form.anyValue && form.num.length > 0 && <div style={styles.orCircle}>OR</div>}
+              </div>}
+              {!form.anyValue && <div style={{minHeight: '10rem'}}>
+                {this.renderNumericalAttributes()}
+                {this.renderCategoricalAttributes()}
+              </div>}
+            </div>
+          }
+          <CalculateFooter addButtonText='ADD THIS'
+                           addFn={() => this.addParameterToSearchItem()}
+                           backFn={() => back()}
+                           calculateFn={() => this.requestPreview()}
+                           calculating={calculating}
+                           count={count}
+                           disableAdd={this.disableAddButton}
+                           disableCalculate={this.disableCalculateButton}/>
         </div>
       );
     }
