+++ ui/src/app/cohort-search/cohort-page/cohort-page.component.spec.tsx
@@ -2,10 +2,9 @@ import {mount} from 'enzyme';
 import * as React from 'react';
 
 import {registerApiClient} from 'app/services/swagger-fetch-clients';
-import {currentWorkspaceStore, queryParamsStore, serverConfigStore} from 'app/utils/navigation';
+import {currentWorkspaceStore, queryParamsStore} from 'app/utils/navigation';
 import {cdrVersionStore} from 'app/utils/navigation';
 import {CohortBuilderApi, CohortsApi} from 'generated/fetch';
-import defaultServerConfig from 'testing/default-server-config';
 import {waitOneTickAndUpdate} from 'testing/react-test-helpers';
 import {cdrVersionListResponse} from 'testing/stubs/cdr-versions-api-stub';
 import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
@@ -19,10 +18,6 @@ describe('CohortPage', () => {
     cdrVersionStore.next(cdrVersionListResponse);
     registerApiClient(CohortBuilderApi, new CohortBuilderServiceStub());
     registerApiClient(CohortsApi, new CohortsApiStub());
-    serverConfigStore.next({
-      ...defaultServerConfig,
-      enableCohortBuilderV2: false
-    });
   });
 
   it('should render', () => {
