+++ ui/src/app/cohort-search/tree-node/tree-node.component.spec.tsx
@@ -2,9 +2,8 @@ import {mount} from 'enzyme';
 import * as React from 'react';
 
 import {registerApiClient} from 'app/services/swagger-fetch-clients';
-import {currentWorkspaceStore, serverConfigStore} from 'app/utils/navigation';
+import {currentWorkspaceStore,} from 'app/utils/navigation';
 import {CohortBuilderApi, Domain} from 'generated/fetch';
-import defaultServerConfig from 'testing/default-server-config';
 import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
 import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
 import {NodeProp, TreeNode} from './tree-node.component';
@@ -48,10 +47,6 @@ describe('TreeNode', () => {
     currentWorkspaceStore.next({
       ...workspaceDataStub,
       cdrVersionId: '1',
-    });
-    serverConfigStore.next({
-      ...defaultServerConfig,
-      enableCohortBuilderV2: false
     });
   });
   it('should create', () => {
